requirements.txt
aioredis==1.3.1
aliyun-python-sdk-core==2.13.26
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-kms==2.12.0
amqp==2.5.2
anyjson==0.3.3
asgiref==3.2.8
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.1
Automat==20.2.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
crcmod==1.7
cryptography==2.9.2
daphne==2.5.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2.12
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-formtools==2.2
django-haystack==2.8.1
django-import-export==2.4.0
django-redis==4.12.1
django-reversion==3.0.8
django-simpleui==2020.9.26
django-timezone-field==4.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
dnspython==1.16.0
drf-extensions==0.6.0
drf-haystack==1.8.6
elasticsearch==2.0.0
et-xmlfile==1.0.1
eventlet==0.25.2
flower==0.9.4
future==0.18.2
greenlet==0.4.15
hiredis==1.0.1
httplib2==0.9.2
humanize==0.5.1
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.0
incremental==17.5.0
jdcal==1.4.1
jieba==0.42.1
jmespath==0.10.0
kombu==4.6.8
lxml==4.5.2
MarkupPy==1.14
monotonic==1.5
msgpack==0.6.2
mysqlclient==1.4.6
numpy==1.18.4
odfpy==1.4.1
openpyxl==3.0.5
oss2==2.13.0
Pillow==7.1.2
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.8
pyDes==2.0.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pypiwin32==223
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
pywin32==228
PyYAML==5.3.1
qiniu==7.2.9
redis==3.5.2
requests==2.23.0
retry==0.9.2
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
supervisor==4.2.1
tablib==2.0.0
tornado==6.0.4
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.9
vine==1.3.0
xadmin2==2.0.3
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
zope.interface==5.1.0
